backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pillow>=10.0.0

# AI and API dependencies
openai>=1.3.0
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0